+++ kibot/bom/bom.py
@@ -274,7 +274,7 @@ class ComponentGroup(object):
             for n in self.components:
                 if n.project == sch.name:
                     S.add(n.ref_id+n.ref_prefix, _suffix_to_num(n.ref_suffix))
-            result = S.flush(self.cfg.ref_separator)
+            result = S.flush(self.cfg.ref_separator, self.cfg.ref_range_separator)
             if result:
                 if refs:
                     refs += self.cfg.ref_separator